refactor: update imports to use new calendar and can-do-list types
