+++ frontend/app/dashboard/calendar/types/eventHooks.ts
@@ -1,4 +1,4 @@
-import { CalendarEvent, EventFormValues, Calendar } from '@/utils/types';
+import { CalendarEvent, EventFormValues, Calendar } from '@/utils/calendar/calendar-types';
 
 export interface EventState {
   events: CalendarEvent[];
