+++ frontend/components/dashboard/calendar/calendar-grid.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState, useRef } from 'react';
 
 import { format, isSameDay, differenceInMinutes } from 'date-fns';
 
-import { CalendarEvent, RecurrenceFrequency } from '@/utils/types';
+import { CalendarEvent, RecurrenceFrequency } from '@/utils/calendar/calendar-types';
 import { generateTimeSlots } from '@/utils/calendar/calendar';
 import { calculateCalendarDimensions, calculateEventRendering, groupOverlappingEvents } from '@/utils/calendar/calendar-render';
 import {
