+++ frontend/utils/calendar/calendar-render.ts
@@ -1,5 +1,5 @@
 import { format, isSameDay, differenceInMinutes, setHours, setMinutes } from 'date-fns';
-import { CalendarEvent } from '../types';
+import { CalendarEvent } from './calendar-types';
 
 export interface CalendarEventRendering {
   eventStyles: {
