+++ frontend/utils/calendar/calendar-types.ts
@@ -1,25 +1,3 @@
-export interface CanDoItem {
-  id: string;
-  content: string;
-  completed: boolean;
-  createdAt: Date;
-  updatedAt?: Date;
-}
-
-export interface EncryptedCanDoItem {
-  id: string;
-  user_id: string;
-  encrypted_data: string;
-  iv: string;
-  salt: string;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface CanDoList {
-  items: CanDoItem[];
-}
-
 // Define recurring event frequency options
 export enum RecurrenceFrequency {
   None = 'none',
